+++ chocolatey-visualstudio.extension/extensions/Get-VSChannelReference.ps1
@@ -3,7 +3,7 @@ function Get-VSChannelReference
     [CmdletBinding()]
     Param
     (
-        [Parameter(Mandatory = $true)] [ValidateSet('2017', '2019')] [string] $VisualStudioYear,
+        [Parameter(Mandatory = $true)] [ValidateSet('2017', '2019', '2022')] [string] $VisualStudioYear,
         [bool] $Preview
     )
 
@@ -11,6 +11,7 @@ function Get-VSChannelReference
     {
         '2017' { $majorVersion = 15 }
         '2019' { $majorVersion = 16 }
+        '2022' { $majorVersion = 17 }
         default { throw "Unsupported VisualStudioYear: $VisualStudioYear"}
     }